+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Added
+- rake task for generating reports
+
 ## [0.1.3] - 2023-09-01
 ### Modified
 - scraping ruby files with many attributes per line