+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 ### Added
 - rake task for generating reports
+- rake task for health check with reporting to console
 
 ## [0.1.3] - 2023-09-01
 ### Modified