+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## [0.2.0] - 2024-09-13
 ### Added
 - rake task for generating reports
 - rake task for health check with reporting to console